Excluded the downloadLosses endpoint from the accept=json header rule to enable proper Swagger documentation.

+++ app/database/migrations/2022_12_17_194359_add_date_removed_to_basket_product.php
@@ -26,9 +26,12 @@ return new class extends Migration
 	 */
 	public function down()
 	{
-		Schema::table('basket_product', function (Blueprint $table)
+		if (Schema::hasColumn('basket_product', 'date_removed'))
 		{
-			$table->dropColumn('date_removed');
-		});
+			Schema::table('basket_product', function (Blueprint $table)
+			{
+				$table->dropColumn('date_removed');
+			});
+		}
 	}
 };
